Simplify usage: get rid of the check CLI command. Fix report layout of the health check.
