+++ README.md
@@ -16,24 +16,29 @@ $ go get github.com/adyatlov/bun/bun
 $ mkdir bundle
 $ unzip bundle.zip -d bundle
 $ cd bundle
-bundle$ bun check
-PROBLEM: dcos-version - Versions are different
-Details:
-master 172.20.0.22 has DC/OS version 1.11.0
-master 172.20.0.24 has DC/OS version 1.11.0
-agent 172.20.0.25 has DC/OS version 1.11.0
-public agent 172.20.0.26 has DC/OS version 1.11.0
-agent 172.20.0.27 has DC/OS version 1.11.0
-agent 172.20.0.29 has DC/OS version 1.11.0
-agent 172.20.0.21 has DC/OS version 1.10.1
+bundle$ bun 
+[PROBLEM] "dcos-version" - Versions are different.
+--------
+Problems
+--------
 master 172.20.0.23 has DC/OS version 1.11.0
+master 172.20.0.24 has DC/OS version 1.11.0
+agent 172.20.0.21 has DC/OS version 1.10.1
+agent 172.20.0.25 has DC/OS version 1.11.0
+agent 172.20.0.27 has DC/OS version 1.11.0
 agent 172.20.0.28 has DC/OS version 1.11.0
+agent 172.20.0.29 has DC/OS version 1.11.0
+public agent 172.20.0.26 has DC/OS version 1.11.0
 
-PROBLEM: health - Some DC/OS systemd units are not healthy.
-Details:
-172.20.0.21 dcos-docker-gc.service: health = 1
+[PROBLEM] "health" - Problems were found.
+--------
+Problems
+--------
+agent 172.20.0.21: The following components are not healthy:
+dcos-docker-gc.service: health = 1
 
-OK: node-count - Masters: 3, Agents: 5, Public Agents: 1, Total: 9
+[OK] "mesos-actor-mailboxes" - All Mesos actors are fine.
+[OK] "node-count" - Masters: 3, Agents: 5, Public Agents: 1, Total: 9
 ```
 
 ## Feedback
